# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend API for EndGit services.",
        "title": "EndGit API",
        "contact": {},
        "version": "0.1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/builds/{id}": {
            "get": {
                "description": "Returns details for a build by ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "builds"
                ],
                "summary": "Build details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Build ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.BuildDetailsResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "description": "Returns service health status and runtime metadata.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Service health check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.HealthResponse"
                        }
                    }
                }
            }
        },
        "/plugins": {
            "get": {
                "description": "Returns a paginated list of plugins.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "plugins"
                ],
                "summary": "List plugins",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Number of items (default 20, max 100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Pagination offset (default 0)",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.ListPluginsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/plugins/{slug}": {
            "get": {
                "description": "Returns details for a plugin by slug.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "plugins"
                ],
                "summary": "Plugin details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Plugin slug",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.PluginDetailsResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handlers.BuildDetailsResponse": {
            "type": "object",
            "properties": {
                "artifactSize": {
                    "type": "integer"
                },
                "artifactSizeLinux": {
                    "type": "integer"
                },
                "artifactSizeWin": {
                    "type": "integer"
                },
                "artifactUrl": {
                    "type": "string"
                },
                "artifactUrlLinux": {
                    "type": "string"
                },
                "artifactUrlWin": {
                    "type": "string"
                },
                "branch": {
                    "type": "string"
                },
                "buildNumber": {
                    "type": "integer"
                },
                "commitHash": {
                    "type": "string"
                },
                "commitMessage": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "duration": {
                    "type": "integer"
                },
                "finishedAt": {
                    "type": "string"
                },
                "ghActionsRunId": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "isRelease": {
                    "type": "boolean"
                },
                "linuxBuildStatus": {
                    "type": "string"
                },
                "logs": {
                    "type": "string"
                },
                "pluginId": {
                    "type": "string"
                },
                "safeScore": {
                    "type": "integer"
                },
                "scanResults": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "triggerType": {
                    "type": "string"
                },
                "winBuildStatus": {
                    "type": "string"
                }
            }
        },
        "handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "handlers.HealthResponse": {
            "type": "object",
            "properties": {
                "startedAt": {
                    "type": "string"
                },
                "status": {
                    "type": "string",
                    "example": "ok"
                },
                "timestamp": {
                    "type": "string"
                },
                "uptime": {
                    "type": "string",
                    "example": "1h2m3s"
                },
                "uptimeSeconds": {
                    "type": "integer",
                    "example": 3723
                }
            }
        },
        "handlers.ListPluginsResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.PluginListItem"
                    }
                },
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                }
            }
        },
        "handlers.PluginDetailsResponse": {
            "type": "object",
            "properties": {
                "authorId": {
                    "type": "string"
                },
                "commentCount": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "displayName": {
                    "type": "string"
                },
                "downloads": {
                    "type": "integer"
                },
                "iconUrl": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "isFeatured": {
                    "type": "boolean"
                },
                "isVerified": {
                    "type": "boolean"
                },
                "keywords": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "license": {
                    "type": "string"
                },
                "longDescription": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "pluginType": {
                    "type": "string"
                },
                "qualityBadge": {
                    "type": "string"
                },
                "repoUrl": {
                    "type": "string"
                },
                "reviewBuildId": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "stars": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "statusReason": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "trustScore": {
                    "type": "number"
                },
                "updatedAt": {
                    "type": "string"
                },
                "webhookId": {
                    "type": "string"
                }
            }
        },
        "handlers.PluginListItem": {
            "type": "object",
            "properties": {
                "authorId": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "displayName": {
                    "type": "string"
                },
                "downloads": {
                    "type": "integer"
                },
                "iconUrl": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "isFeatured": {
                    "type": "boolean"
                },
                "isVerified": {
                    "type": "boolean"
                },
                "keywords": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                },
                "pluginType": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "stars": {
                    "type": "integer"
                }
            }
        }
    }
}